kernel-3.0.x: implement upstream patches from 3.4.98-102

+++ trunk/linux-3.0.x/include/linux/slab.h
@@ -222,7 +222,7 @@ size_t ksize(const void *);
  */
 static inline void *kcalloc(size_t n, size_t size, gfp_t flags)
 {
-	if (size != 0 && n > ULONG_MAX / size)
+	if (size != 0 && n > SIZE_MAX / size)
 		return NULL;
 	return __kmalloc(n * size, flags | __GFP_ZERO);
 }
